uint8 optimizations

+++ src/trigramAnalysis.nim
@@ -1,5 +1,5 @@
 # Basic pattern checks
-proc isSameHandTri(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameHandTri(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   let
     f1 = getFinger(map, row0, col0)
     f2 = getFinger(map, row1, col1)
@@ -12,7 +12,7 @@ proc isSameColTri(row0, col0, row1, col1, row2, col2: int): bool =
 proc isSameRowTri(row0, col0, row1, col1, row2, col2: int): bool =
   row0 == row1 and row1 == row2
 
-proc isSameRowModTri(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowModTri(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   row0 == row1 and row1 == row2 and
   not isStretch(map, row0, col0) and
   not isStretch(map, row1, col1) and
@@ -22,7 +22,7 @@ proc isSamePosTri(row0, col0, row1, col1, row2, col2: int): bool =
   isSameColTri(row0, col0, row1, col1, row2, col2) and
   isSameRowTri(row0, col0, row1, col1, row2, col2)
 
-proc isAdjacentFingerTri(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isAdjacentFingerTri(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   let
     f1 = getFinger(map, row0, col0)
     f2 = getFinger(map, row1, col1)
@@ -38,7 +38,7 @@ proc isAdjacentFingerTri(map: FingerMap, row0, col0, row1, col1, row2, col2: int
   isAdjacent(map, f2, f3) and
   f1 != f2 and f2 != f3 and f1 != f3
 
-proc isSameFingerTri(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameFingerTri(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   let
     f1 = getFinger(map, row0, col0)
     f2 = getFinger(map, row1, col1)
@@ -48,7 +48,7 @@ proc isSameFingerTri(map: FingerMap, row0, col0, row1, col1, row2, col2: int): b
   not isSamePosBi(row0, col0, row1, col1) and
   not isSamePosBi(row1, col1, row2, col2)
 
-proc isRedirect(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isRedirect(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   let
     f1 = getFinger(map, row0, col0)
     f2 = getFinger(map, row1, col1)
@@ -63,7 +63,7 @@ proc isRedirect(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
   ((ord(f1.get) < ord(f2.get) and ord(f2.get) > ord(f3.get)) or
    (ord(f1.get) > ord(f2.get) and ord(f2.get) < ord(f3.get)))
 
-proc isBadRedirect(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isBadRedirect(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   let
     f1 = getFinger(map, row0, col0)
     f2 = getFinger(map, row1, col1)
@@ -78,13 +78,13 @@ proc isBadRedirect(map: FingerMap, row0, col0, row1, col1, row2, col2: int): boo
   f3.get notin {LI, RI}
 
 # Alternation patterns
-proc isAlt(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isAlt(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   not isSameHandBi(map, row0, col0, row1, col1) and
   not isSameHandBi(map, row1, col1, row2, col2) and
   not isSameFingerBi(map, row0, col0, row2, col2) and
   not isSamePosBi(row0, col0, row2, col2)
 
-proc isRoll(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isRoll(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   (isSameHandBi(map, row0, col0, row1, col1) and
     not isSameHandBi(map, row1, col1, row2, col2) and
     not isSameFingerBi(map, row0, col0, row1, col1) and
@@ -94,7 +94,7 @@ proc isRoll(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
     not isSameFingerBi(map, row1, col1, row2, col2) and
     not isSamePosBi(row1, col1, row2, col2))
 
-proc isRollIn(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isRollIn(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   let
     f1 = getFinger(map, row0, col0)
     f2 = getFinger(map, row1, col1)
@@ -117,56 +117,56 @@ proc isRollIn(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
     getHand(f2) == some(Right) and
     ord(f2.get) > ord(f3.get)))
 
-proc isAltIn(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isAltIn(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isAlt(map, row0, col0, row1, col1, row2, col2) and
   isRollIn(map, row0, col0, row2, col2, row1, col1)
 
-proc isAltOut(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isAltOut(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isAlt(map, row0, col0, row1, col1, row2, col2) and
   not isAltIn(map, row0, col0, row1, col1, row2, col2)
 
 # Same Row patterns
-proc isSameRowAlt(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowAlt(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isAlt(map, row0, col0, row1, col1, row2, col2) and
   isSameRowBi(row0, col0, row2, col2)
 
-proc isSameRowAltIn(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowAltIn(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isSameRowAlt(map, row0, col0, row1, col1, row2, col2) and
   isRollIn(map, row0, col0, row2, col2, row1, col1)
 
-proc isSameRowAltOut(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowAltOut(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isSameRowAlt(map, row0, col0, row1, col1, row2, col2) and
   not isSameRowAltIn(map, row0, col0, row1, col1, row2, col2)
 
 # Adjacent Finger patterns
-proc isAdjacentFingerAlt(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isAdjacentFingerAlt(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isAlt(map, row0, col0, row1, col1, row2, col2) and
   isAdjacentFingerBi(map, row0, col0, row2, col2)
 
-proc isAdjacentFingerAltIn(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isAdjacentFingerAltIn(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isAdjacentFingerAlt(map, row0, col0, row1, col1, row2, col2) and
   isRollIn(map, row0, col0, row2, col2, row1, col1)
 
-proc isAdjacentFingerAltOut(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isAdjacentFingerAltOut(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isAdjacentFingerAlt(map, row0, col0, row1, col1, row2, col2) and
   not isAdjacentFingerAltIn(map, row0, col0, row1, col1, row2, col2)
 
 # Same Row Adjacent Finger patterns
-proc isSameRowAdjacentFingerAlt(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowAdjacentFingerAlt(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isAlt(map, row0, col0, row1, col1, row2, col2) and
   isSameRowBi(row0, col0, row2, col2) and
   isAdjacentFingerBi(map, row0, col0, row2, col2)
 
-proc isSameRowAdjacentFingerAltIn(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowAdjacentFingerAltIn(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isSameRowAdjacentFingerAlt(map, row0, col0, row1, col1, row2, col2) and
   isRollIn(map, row0, col0, row2, col2, row1, col1)
 
-proc isSameRowAdjacentFingerAltOut(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowAdjacentFingerAltOut(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isSameRowAdjacentFingerAlt(map, row0, col0, row1, col1, row2, col2) and
   not isSameRowAdjacentFingerAltIn(map, row0, col0, row1, col1, row2, col2)
 
 # One Hand patterns
-proc isOneHand(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isOneHand(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   let
     f1 = getFinger(map, row0, col0)
     f2 = getFinger(map, row1, col1)
@@ -179,7 +179,7 @@ proc isOneHand(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
   ((ord(f1.get) < ord(f2.get) and ord(f2.get) < ord(f3.get)) or
     (ord(f1.get) > ord(f2.get) and ord(f2.get) > ord(f3.get)))
 
-proc isOneHandIn(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isOneHandIn(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   let
     f1 = getFinger(map, row0, col0)
     f2 = getFinger(map, row1, col1)
@@ -194,94 +194,94 @@ proc isOneHandIn(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool 
     (getHand(f1) == some(Right) and
     ord(f1.get) > ord(f2.get) and ord(f2.get) > ord(f3.get)))
 
-proc isOneHandOut(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isOneHandOut(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isOneHand(map, row0, col0, row1, col1, row2, col2) and
   not isOneHandIn(map, row0, col0, row1, col1, row2, col2)
 
 # Same Row One Hand patterns
-proc isSameRowOneHand(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowOneHand(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isOneHand(map, row0, col0, row1, col1, row2, col2) and
   isSameRowModTri(map, row0, col0, row1, col1, row2, col2)
 
-proc isSameRowOneHandIn(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowOneHandIn(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isOneHandIn(map, row0, col0, row1, col1, row2, col2) and
   isSameRowModTri(map, row0, col0, row1, col1, row2, col2)
 
-proc isSameRowOneHandOut(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowOneHandOut(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isOneHandOut(map, row0, col0, row1, col1, row2, col2) and
   isSameRowModTri(map, row0, col0, row1, col1, row2, col2)
 
 # Adjacent Finger One Hand patterns
-proc isAdjacentFingerOneHand(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isAdjacentFingerOneHand(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isOneHand(map, row0, col0, row1, col1, row2, col2) and
   isAdjacentFingerTri(map, row0, col0, row1, col1, row2, col2)
 
-proc isAdjacentFingerOneHandIn(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isAdjacentFingerOneHandIn(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isOneHandIn(map, row0, col0, row1, col1, row2, col2) and
   isAdjacentFingerTri(map, row0, col0, row1, col1, row2, col2)
 
-proc isAdjacentFingerOneHandOut(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isAdjacentFingerOneHandOut(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isOneHandOut(map, row0, col0, row1, col1, row2, col2) and
   isAdjacentFingerTri(map, row0, col0, row1, col1, row2, col2)
 
 # Same Row Adjacent Finger One Hand patterns
-proc isSameRowAdjacentFingerOneHand(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowAdjacentFingerOneHand(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isOneHand(map, row0, col0, row1, col1, row2, col2) and
   isSameRowModTri(map, row0, col0, row1, col1, row2, col2) and
   isAdjacentFingerTri(map, row0, col0, row1, col1, row2, col2)
 
-proc isSameRowAdjacentFingerOneHandIn(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowAdjacentFingerOneHandIn(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isOneHandIn(map, row0, col0, row1, col1, row2, col2) and
   isSameRowModTri(map, row0, col0, row1, col1, row2, col2) and
   isAdjacentFingerTri(map, row0, col0, row1, col1, row2, col2)
 
-proc isSameRowAdjacentFingerOneHandOut(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowAdjacentFingerOneHandOut(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isOneHandOut(map, row0, col0, row1, col1, row2, col2) and
   isSameRowModTri(map, row0, col0, row1, col1, row2, col2) and
   isAdjacentFingerTri(map, row0, col0, row1, col1, row2, col2)
 
 # Roll patterns
-proc isRollOut(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isRollOut(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isRoll(map, row0, col0, row1, col1, row2, col2) and
   not isRollIn(map, row0, col0, row1, col1, row2, col2)
 
-proc isSameRowRoll(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowRoll(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isRoll(map, row0, col0, row1, col1, row2, col2) and
   ((isSameHandBi(map, row0, col0, row1, col1) and
     isSameRowModBi(map, row0, col0, row1, col1)) or
     (isSameHandBi(map, row1, col1, row2, col2) and
     isSameRowModBi(map, row1, col1, row2, col2)))
 
-proc isSameRowRollIn(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowRollIn(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isRollIn(map, row0, col0, row1, col1, row2, col2) and
   isSameRowRoll(map, row0, col0, row1, col1, row2, col2)
 
-proc isSameRowRollOut(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowRollOut(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isRollOut(map, row0, col0, row1, col1, row2, col2) and
   isSameRowRoll(map, row0, col0, row1, col1, row2, col2)
 
 # Adjacent Finger Roll patterns
-proc isAdjacentFingerRoll(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isAdjacentFingerRoll(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isRoll(map, row0, col0, row1, col1, row2, col2) and
   isAdjacentFingerBi(map, row0, col0, row1, col1)
 
-proc isAdjacentFingerRollIn(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isAdjacentFingerRollIn(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isRollIn(map, row0, col0, row1, col1, row2, col2) and
   isAdjacentFingerRoll(map, row0, col0, row1, col1, row2, col2)
 
-proc isAdjacentFingerRollOut(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isAdjacentFingerRollOut(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isRollOut(map, row0, col0, row1, col1, row2, col2) and
   isAdjacentFingerRoll(map, row0, col0, row1, col1, row2, col2)
 
 # Same Row Adjacent Finger Roll patterns
-proc isSameRowAdjacentFingerRoll(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowAdjacentFingerRoll(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isSameRowRoll(map, row0, col0, row1, col1, row2, col2) and
   isAdjacentFingerRoll(map, row0, col0, row1, col1, row2, col2)
 
-proc isSameRowAdjacentFingerRollIn(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowAdjacentFingerRollIn(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isRollIn(map, row0, col0, row1, col1, row2, col2) and
   isSameRowAdjacentFingerRoll(map, row0, col0, row1, col1, row2, col2)
 
-proc isSameRowAdjacentFingerRollOut(map: FingerMap, row0, col0, row1, col1, row2, col2: int): bool =
+proc isSameRowAdjacentFingerRollOut(map: FingerMap, row0, col0, row1, col1, row2, col2: uint8): bool =
   isRollOut(map, row0, col0, row1, col1, row2, col2) and
   isSameRowAdjacentFingerRoll(map, row0, col0, row1, col1, row2, col2)
